Refactor: Rename configs.py to agent_configs.py and update import paths for better organization
