+++ agents/utils/credentials.py
@@ -1,4 +1,3 @@
-import logging
 import os
 
 from dotenv import load_dotenv
